--- src/app/CoursesSection/page.tsx
@@ -1,11 +1,0 @@
-import CoursesSection from "@/components/coursesSection";
-
-const coursesSectionPage = () => {
-  return (
-    <main className="min-h-screen flex items-center justify-center bg-gray-900">
-      <CoursesSection />
-    </main>
-  );
-};
-
-export default coursesSectionPage;
