--- src/app/FooterPage/page.tsx
@@ -1,5 +1,0 @@
-import Footer from "../components/footer";
-
-export default function FooterPage() {
-  return <Footer />;
-}
