--- src/app/WelcomeSection/page.tsx
@@ -1,11 +1,0 @@
-import WelcomeSection from "@/app/components/WelcomeSection";
-
-const WelcomePage = () => {
-  return (
-    <main className="min-h-screen flex items-center justify-center bg-gray-900">
-      <WelcomeSection />
-    </main>
-  );
-};
-
-export default WelcomePage;
