--- src/app/home/page.tsx
@@ -1,10 +1,0 @@
-import ComponentsPage from '../components/page';
-
-export default function Home() {
-  return (
-    <div>
-      <h1>Home Page proof/h1>
-      <ComponentsPage />
-    </div>
-  );
-}
