+++ src/app/page.tsx
@@ -1,9 +1,19 @@
-import ConnectWallet from "@/app/components/ConnectWalletButton";
+import CoursesSection from '@/components/coursesSection';
+import Footer from '@/components/footer';
+import Navbar from '@/components/nabvar';
+import StellarSection from '@/components/stellarSection';
+import WelcomeSection from '@/components/WelcomeSection';
 
-export default function HomePage() {
+export default function Home() {
   return (
-    <div className="flex items-center justify-center min-h-screen bg-gray-900 text-white">
-      <ConnectWallet />
+    <div className="min-h-screen bg-gray-900 text-white">
+      <div className='mx-[200px]'>
+        <Navbar />
+        <WelcomeSection />
+        <CoursesSection />
+        <StellarSection />
+      </div>
+      <Footer />
     </div>
   );
 }
