+++ src/components/AlbedoIcon.tsx
