+++ src/components/xBullIcon.tsx
